All in on async
