+++ crates/spin-sdk/Cargo.toml
@@ -21,7 +21,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 name = "spin_sdk"
 
 [features]
-default = ["http", "key-value", "json", "llm", "mqtt", "mysql", "mysql-async", "pg", "postgres4-types", "redis", "sqlite", "variables", "export-sdk-language"]
+default = ["http", "key-value", "json", "llm", "mqtt", "mysql", "pg", "postgres4-types", "redis", "sqlite", "variables", "export-sdk-language"]
 export-sdk-language = []
 http = ["dep:bytes", "dep:http-body", "dep:http-body-util", "dep:hyperium"]
 http-middleware = ["http"]
@@ -31,7 +31,6 @@ json = ["dep:serde", "dep:serde_json"]
 llm = []
 mqtt = []
 mysql = []
-mysql-async = []
 pg = ["dep:chrono"]
 postgres4-types = ["pg", "json", "dep:rust_decimal", "dep:uuid", "dep:postgres_range"]
 redis = []
