+++ crates/spin-sdk/src/mysql.rs
@@ -1,8 +1,4 @@
-//! MySQL relational database storage for downlevel runtimes.
-//!
-//! This module contains the synchronous
-//! interface. For new code which does not need to target down-level
-//! versions of Spin, we recommend the [async interface](crate::mysql_async).
+//! MySQL relational database storage.
 //!
 //! You can use the [`Decode`] trait to convert a [`DbValue`] to a
 //! suitable Rust type. The following table shows available conversions.
@@ -25,6 +21,9 @@
 //! | `String`  | str(string)         | VARCHAR, CHAR, TEXT     |
 //! | `Vec<u8>` | binary(list\<u8\>)  | VARBINARY, BINARY, BLOB |
 
+use crate::wit_bindgen;
+use std::sync::Arc;
+
 #[doc(hidden)]
 /// Module containing wit bindgen generated code.
 ///
@@ -35,10 +34,12 @@ pub mod wit {
 
     wit_bindgen::generate!({
         runtime_path: "crate::wit_bindgen::rt",
-        world: "spin-sdk-mysql",
+        world: "spin-sdk-mysql-v3",
         path: "wit",
         generate_all,
     });
+
+    pub use spin::mysql::mysql;
 }
 
 /// An open connection to a MySQL database.
@@ -50,21 +51,21 @@ pub mod wit {
 /// ```no_run
 /// use spin_sdk::mysql::{Connection, Decode, ParameterValue};
 ///
-/// # fn run() -> anyhow::Result<()> {
+/// # async fn run() -> anyhow::Result<()> {
 /// # let min_age = 0;
-/// let db = Connection::open("mysql://root:my_password@localhost/mydb")?;
+/// let db = Connection::open("mysql://root:my_password@localhost/mydb").await?;
 ///
-/// let query_result = db.query(
-///     "SELECT * FROM users WHERE age < ?",
-///     &[ParameterValue::Int32(20)]
-/// )?;
+/// let mut query_result = db.query(
+///     "SELECT * FROM users WHERE age >= ?",
+///     &[min_age.into()]
+/// ).await?;
 ///
-/// let name_index = query_result.columns.iter().position(|c| c.name == "name").unwrap();
-///
-/// for row in &query_result.rows {
-///     let name = String::decode(&row[name_index])?;
+/// while let Some(row) = query_result.next().await {
+///     let name = row.get::<String>("name").unwrap();
 ///     println!("Found user {name}");
 /// }
+///
+/// query_result.result().await?;
 /// # Ok(())
 /// # }
 /// ```
@@ -75,16 +76,19 @@ pub mod wit {
 /// ```no_run
 /// use spin_sdk::mysql::{Connection, Decode};
 ///
-/// # fn run() -> anyhow::Result<()> {
-/// let db = Connection::open("mysql://root:my_password@localhost/mydb")?;
+/// # async fn run() -> anyhow::Result<()> {
+/// let db = Connection::open("mysql://root:my_password@localhost/mydb").await?;
 ///
-/// let query_result = db.query("SELECT COUNT(*) FROM users", &[])?;
+/// let mut query_result = db.query("SELECT COUNT(*) FROM users", &[]).await?;
 ///
-/// assert_eq!(1, query_result.columns.len());
-/// assert_eq!("COUNT(*)", query_result.columns[0].name);
-/// assert_eq!(1, query_result.rows.len());
+/// assert_eq!(1, query_result.columns().len());
+/// assert_eq!("COUNT(*)", query_result.columns()[0].name);
 ///
-/// let count = i64::decode(&query_result.rows[0][0])?;
+/// let rows = query_result.collect().await?;
+///
+/// assert_eq!(1, rows.len());
+///
+/// let count = &rows[0][0];
 /// # Ok(())
 /// # }
 /// ```
@@ -95,57 +99,199 @@ pub mod wit {
 /// ```no_run
 /// use spin_sdk::mysql::{Connection, ParameterValue};
 ///
-/// # fn run() -> anyhow::Result<()> {
-/// let db = Connection::open("mysql://root:my_password@localhost/mydb")?;
+/// # async fn run() -> anyhow::Result<()> {
+/// let db = Connection::open("mysql://root:my_password@localhost/mydb").await?;
 ///
-/// let rows_affected = db.execute(
+/// db.execute(
 ///     "DELETE FROM users WHERE name = ?",
-///     &[ParameterValue::Str("Baldrick".to_owned())]
-/// )?;
+///     &["Baldrick".to_owned().into()]
+/// ).await?;
 /// # Ok(())
 /// # }
 /// ```
-#[doc(inline)]
-pub use wit::fermyon::spin::mysql::Connection;
+pub struct Connection(wit::mysql::Connection);
 
-/// The result of a database query.
-///
-/// # Examples
-///
-/// Load a set of rows from a local MySQL database, and iterate over them
-/// selecting one field from each. The columns collection allows you to find
-/// column indexes for column names; you can bypass this lookup if you name
-/// specific columns in the query.
-///
-/// ```no_run
-/// use spin_sdk::mysql::{Connection, Decode, ParameterValue};
-///
-/// # fn run() -> anyhow::Result<()> {
-/// # let min_age = 0;
-/// let db = Connection::open("mysql://root:my_password@localhost/mydb")?;
-///
-/// let query_result = db.query(
-///     "SELECT * FROM users WHERE age >= ?",
-///     &[ParameterValue::Int32(min_age)]
-/// )?;
-///
-/// let name_index = query_result.columns.iter().position(|c| c.name == "name").unwrap();
-///
-/// for row in &query_result.rows {
-///     let name = String::decode(&row[name_index])?;
-///     println!("Found user {name}");
-/// }
-/// # Ok(())
-/// # }
-/// ```
-#[doc(inline)]
-pub use wit::fermyon::spin::mysql::RowSet;
+impl Connection {
+    /// Open a connection to a MySQL database.
+    ///
+    /// The address may be in connection string form (`"host=... dbname=..."`)
+    /// or in URL form (`"mysql://<host>/<dbname>?..."`).
+    pub async fn open(address: impl Into<String>) -> Result<Self, Error> {
+        let inner = wit::mysql::Connection::open(address.into()).await?;
+        Ok(Self(inner))
+    }
+
+    /// Query the database.
+    ///
+    /// Use this function for queries that return rows (typically `SELECT` queries).
+    /// For side-effectful queries, see [`Connection::execute`].
+    pub async fn query(
+        &self,
+        statement: impl Into<String>,
+        params: impl Into<Vec<ParameterValue>>,
+    ) -> Result<QueryResult, Error> {
+        let (columns, rows, result) = self.0.query(statement.into(), params.into()).await?;
+        Ok(QueryResult {
+            columns: Arc::new(columns),
+            rows,
+            result,
+        })
+    }
+
+    /// Execute a command against the database.
+    ///
+    /// Use this function for side-effectful queries (such as `INSERT` or `DELETE` queries).
+    /// For queries that return row data, see [`Connection::query`].
+    pub async fn execute(
+        &self,
+        statement: impl Into<String>,
+        params: impl Into<Vec<ParameterValue>>,
+    ) -> Result<(), Error> {
+        self.0
+            .execute(statement.into(), params.into())
+            .await
+            .map_err(Error::MysqlError)
+    }
+}
 
 #[doc(inline)]
-pub use wit::fermyon::spin::mysql::Error as MysqlError;
+pub use wit::mysql::Error as MysqlError;
 
 #[doc(inline)]
-pub use wit::fermyon::spin::rdbms_types::*;
+pub use wit::mysql::{Column, DbDataType, DbValue, ParameterValue};
+
+/// The result of a [`Connection::query`] operation.
+pub struct QueryResult {
+    columns: Arc<Vec<Column>>,
+    rows: wit_bindgen::StreamReader<Vec<DbValue>>,
+    result: wit_bindgen::FutureReader<Result<(), MysqlError>>,
+}
+
+impl QueryResult {
+    /// The columns in the query result.
+    pub fn columns(&self) -> &[Column] {
+        &self.columns
+    }
+
+    /// Gets the next row in the result set.
+    ///
+    /// If this is `None`, there are no more rows available. You _must_
+    /// await [`QueryResult::result()`] to determine if all rows
+    /// were read successfully.
+    pub async fn next(&mut self) -> Option<Row> {
+        self.rows.next().await.map(|r| Row {
+            columns: self.columns.clone(),
+            result: r,
+        })
+    }
+
+    /// Whether the query completed successfully or with an error.
+    pub async fn result(self) -> Result<(), Error> {
+        self.result.await.map_err(Error::MysqlError)
+    }
+
+    /// Collect all rows in the result set.
+    ///
+    /// This is provided for when the result set is small enough to fit in
+    /// memory and you do not require streaming behaviour.
+    pub async fn collect(mut self) -> Result<Vec<Row>, Error> {
+        let mut rows = vec![];
+        while let Some(row) = self.next().await {
+            rows.push(row);
+        }
+        self.result.await.map_err(Error::MysqlError)?;
+        Ok(rows)
+    }
+
+    /// An asynchronous reader for the rows of the query result. Call
+    /// `.next().await` to iterate over the rows. When this returns `None`,
+    /// you have read all available rows. At this point you _must_ check
+    /// [`QueryResult::result()`] to determine if the read completed
+    /// successfully.
+    ///
+    /// This provides each row as a plain vector of database values.
+    /// [`QueryResult::next()`] provides a more ergonomic wrapper.
+    ///
+    /// To collect all rows into a vector, see [`QueryResult::collect`].
+    pub fn rows(&mut self) -> &mut wit_bindgen::StreamReader<Vec<DbValue>> {
+        &mut self.rows
+    }
+
+    /// Extracts the underlying Wasm Component Model results of the query.
+    #[allow(
+        clippy::type_complexity,
+        reason = "sorry clippy that's just what the inner bits are"
+    )]
+    pub fn into_inner(
+        self,
+    ) -> (
+        Vec<Column>,
+        wit_bindgen::StreamReader<Vec<DbValue>>,
+        wit_bindgen::FutureReader<Result<(), MysqlError>>,
+    ) {
+        ((*self.columns).clone(), self.rows, self.result)
+    }
+}
+
+/// A database row result.
+///
+/// There are two representations of a MySQL row in the SDK.  This type is useful for
+/// addressing elements by column name, and is obtained from the [QueryResult::next()] function.
+/// The [DbValue] vector representation is obtained from the [QueryResult::rows()] function, and provides
+/// index-based lookup or low-level access to row values via a vector.
+pub struct Row {
+    columns: Arc<Vec<wit::mysql::Column>>,
+    result: Vec<DbValue>,
+}
+
+impl Row {
+    /// Get a value by its column name. The value is converted to the target type as per the
+    /// conversion table shown in the module documentation.
+    ///
+    /// This function returns None for both no such column _and_ failed conversion. You should use
+    /// it only if you do not need to address errors (that is, if you know that conversion should
+    /// never fail). If your code does not know the type in advance, use the raw [QueryResult::rows()] function
+    /// instead of the [`QueryResult::next()`] or [`QueryResult::collect()`] wrappers to access
+    /// the underlying [DbValue] enum: this will allow you to
+    /// determine the type and process it accordingly.
+    ///
+    /// Additionally, this function performs a name lookup each time it is called. If you are iterating
+    /// over a large number of rows, it's more efficient to use column indexes, either calculated or
+    /// statically known from the column order in the SQL.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use spin_sdk::mysql::{Connection, DbValue};
+    ///
+    /// # async fn run() -> anyhow::Result<()> {
+    /// # let user_id = 0;
+    /// let db = Connection::open("mysql://root:my_password@localhost/mydb").await?;
+    /// let mut query_result = db.query(
+    ///     "SELECT * FROM users WHERE id = ?",
+    ///     &[user_id.into()]
+    /// ).await?;
+    /// let user_row = query_result.next().await.unwrap();
+    ///
+    /// let name = user_row.get::<String>("name").unwrap();
+    /// let age = user_row.get::<i16>("age").unwrap();
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn get<T: Decode>(&self, column: &str) -> Option<T> {
+        let i = self.columns.iter().position(|c| c.name == column)?;
+        let db_value = self.result.get(i)?;
+        Decode::decode(db_value).ok()
+    }
+}
+
+impl std::ops::Index<usize> for Row {
+    type Output = DbValue;
+
+    fn index(&self, index: usize) -> &Self::Output {
+        &self.result[index]
+    }
+}
 
 /// A MySQL error
 #[derive(Debug, thiserror::Error)]
@@ -301,6 +447,30 @@ impl Decode for String {
             ))),
         }
     }
+}
+
+macro_rules! impl_parameter_value_conversions {
+    ($($ty:ty => $id:ident),*) => {
+        $(
+            impl From<$ty> for ParameterValue {
+                fn from(v: $ty) -> ParameterValue {
+                    ParameterValue::$id(v)
+                }
+            }
+        )*
+    };
+}
+
+impl_parameter_value_conversions! {
+    i8 => Int8,
+    i16 => Int16,
+    i32 => Int32,
+    i64 => Int64,
+    f32 => Floating32,
+    f64 => Floating64,
+    bool => Boolean,
+    String => Str,
+    Vec<u8> => Binary
 }
 
 fn format_decode_err(types: &str, value: &DbValue) -> String {
